Upd: Jacob Solomons 1855-1898
I am starting with the earliest generation of the Wiki Tree profiles of
my family and trying to gettheir sources in order as I go. I have the
Wiki Tree bio in a file named such as
jacob_solomons_18550000_18980605.wiki and have moved the sources I
associate with that person into a file with src between the name and
date such as jacob_solomons_src_18550000_18980605.wiki

I am making as little changes as possible as I go along so I can strike
a balance between getting public profiles updated and my own house in
order.